# pixela-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pixela API Documentation",
        "description": "API Documentation for Pixela - A platform for tracking and managing your movie and series watching history",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/series/{seriesId}": {
            "get": {
                "tags": [
                    "Series"
                ],
                "summary": "Get TV series details",
                "description": "Retrieves detailed information about a specific TV series by its ID",
                "operationId": "getSeriesDetails",
                "parameters": [
                    {
                        "name": "seriesId",
                        "in": "path",
                        "description": "ID of the TV series",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "TV series details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SeriesDetails"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "TV series not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/series/trending": {
            "get": {
                "tags": [
                    "Series"
                ],
                "summary": "Get trending TV series",
                "description": "Returns a paginated list of trending TV series",
                "operationId": "getTrendingSeries",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (default: 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of trending TV series",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedSeriesResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/series/top-rated": {
            "get": {
                "tags": [
                    "Series"
                ],
                "summary": "Get top-rated TV series",
                "description": "Returns a paginated list of top-rated TV series",
                "operationId": "getTopRatedSeries",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (default: 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of top-rated TV series",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedSeriesResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/series/on-the-air": {
            "get": {
                "tags": [
                    "Series"
                ],
                "summary": "Get TV series currently on air",
                "description": "Returns a paginated list of TV series currently airing",
                "operationId": "getSeriesOnTheAir",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (default: 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of TV series currently on air",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedSeriesResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/series/discover": {
            "get": {
                "tags": [
                    "Series"
                ],
                "summary": "Discover TV series",
                "description": "Returns a paginated list of discovered TV series",
                "operationId": "getDiscoveredSeries",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (default: 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of discovered TV series",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedSeriesResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/series/genre/{genreId}": {
            "get": {
                "tags": [
                    "Series"
                ],
                "summary": "Get TV series by genre",
                "description": "Returns a paginated list of TV series for a specific genre",
                "operationId": "getSeriesByGenre",
                "parameters": [
                    {
                        "name": "genreId",
                        "in": "path",
                        "description": "ID of the genre",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (default: 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of TV series for the specified genre",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedSeriesResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/series/{seriesId}/cast": {
            "get": {
                "tags": [
                    "Series"
                ],
                "summary": "Get TV series cast",
                "description": "Returns the cast list of a specific TV series",
                "operationId": "getSeriesCast",
                "parameters": [
                    {
                        "name": "seriesId",
                        "in": "path",
                        "description": "ID of the TV series",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "TV series cast",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SeriesCastResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "TV series not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "description": "General error response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "An error occurred"
                    }
                },
                "type": "object"
            },
...
            "PaginatedSeriesResponse": {
                "description": "Paginated response of TV series",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "total_pages": {
                        "type": "integer",
                        "example": 100
                    },
                    "total_results": {
                        "type": "integer",
                        "example": 2000
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SeriesItem"
                        }
                    }
                },
                "type": "object"
            },
...
            "SeriesDetails": {
                "description": "Complete TV series details",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/SeriesItem"
                    },
                    {
                        "properties": {
                            "number_of_seasons": {
                                "type": "integer",
                                "example": 8
                            },
                            "number_of_episodes": {
                                "type": "integer",
                                "example": 73
                            },
                            "status": {
                                "type": "string",
                                "example": "Ended"
                            },
                            "genres": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Authentication using Laravel Sanctum. Include the token in the Authorization header as 'Bearer {token}'",
                "scheme": "bearer"
            }
        }
    },
...
}